requirements.txt
Flask==3.0.3
Werkzeug==3.0.2
PyPDF2==3.0.1
python-docx==0.8.11
